pyproject.toml
[tool.poetry]
name = "newsletter_gen"
version = "0.1.0"
description = "newsletter-gen using crewAI"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = { extras = ["tools"], version = ">=0.67.1,<1.0.0" }
exa-py = "^1.0.9"
langchain = "0.2.16"  # Atualize para uma versão compatível
langchain-core = "0.2.38"  # Ajuste para corresponder à nova versão do langchain
langchain-anthropic = "^0.1.11"
langchain-groq = "^0.1.3"
streamlit = "^1.34.0"
langchain-google-genai = "^1.0.3"
pillow = "^10.3.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
newsletter_gen = "newsletter_gen.main:run"
run_crew = "newsletter_gen.main:run"
train = "newsletter_gen.main:train"
replay = "newsletter_gen.main:replay"
test = "newsletter_gen.main:test"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
